[feat] add CustomExceptionHandler

+++ src/main/java/com/mouthird/devcovid19api/service/NewsService.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.data.domain.PageRequest;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -21,7 +22,17 @@ public class NewsService {
 
     public void addNews(List<News> newsList) { newsRepository.saveAll(newsList); }
 
-    public void putNews(News news) { newsRepository.save(news); }
+    public void putNews(News news) {
+        boolean exists = newsRepository.existsById(news.getId());
+        if (!exists) {
+            throw new IllegalStateException(
+                    "News with id " + news.getId() + " does not exists"
+            );
+        }
+        if(news.getCrawlTime() == null)
+            news.setCrawlTime(LocalDate.now());
+        newsRepository.save(news);
+    }
 
     public List<News> getNews(int limit) {
         PageRequest pageable = PageRequest.of(0, limit, Sort.by(Sort.Direction.DESC, "newsTime"));
@@ -29,6 +40,12 @@ public class NewsService {
     }
 
     public void deleteById(Long Id) {
+        boolean exists = newsRepository.existsById(Id);
+        if (!exists) {
+            throw new IllegalStateException(
+                    "News with id " + Id + " does not exists"
+            );
+        }
         newsRepository.deleteById(Id);
     }
